Restructure (#2)
* Switch over to using the src/ layout

* Add tests with fixes

* Configure GitHub Actions

* Update pyproject.toml

* Update README.rst

+++ src/filib/oanda.py
@@ -278,7 +278,7 @@ class Oanda:
                 pd.to_numeric, errors='ignore')
             trades = json_normalize(account.trades).apply(
                 pd.to_numeric, errors='ignore')
-        orders = pd.Series(name=self.name)
+        orders = pd.Series(name=self.name, dtype=int)
         for instrument in self.instruments:
             base, quote = instrument.split('_')
             params = {'instruments': instrument}
